requirements.txt
google_api_python_client==2.64.0
google_auth_oauthlib==0.5.2
gTTS==2.2.4
mutagen==1.45.1
oauth2client==4.1.3
praw==7.6.0
protobuf==4.21.7
python-dotenv==0.21.0
requests==2.28.1
selenium==4.5.0
Wand==0.6.10
webdriver_manager==3.8.3
youtube_dl==2021.12.17
